refactor : 폴더 구조 변경
- /db => json 파일 폴더
- /public => html, css, img, js 를 포함하는 폴더
    /html => html 파일 폴더
    /css => css 파일 폴더
    /img => 이미지 파일 폴더
    /js => 클라이언트 단 js 파일 폴더
- /test => test 관련 파일 폴더
- / => server 관련 파일
