--- data/todos.json
@@ -1,19 +1,0 @@
-{
-	"진진진": [
-		{
-			"todos_id": 5127,
-			"todos_status": "todo",
-			"todos_contents": "wls"
-		},
-		{
-			"todos_id": 7064,
-			"todos_status": "todo",
-			"todos_contents": "진님"
-		},
-		{
-			"todos_id": 6069,
-			"todos_status": "todo",
-			"todos_contents": "진!!!!"
-		}
-	]
-}
