+++ public/error-500.html
@@ -5,11 +5,11 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<meta http-equiv="X-UA-Compatible" content="ie=edge" />
 		<title>500 ERROR</title>
-		<link rel="shortcut icon" href="css/img/favicon.ico" />
+		<link rel="shortcut icon" href="img/favicon.ico" />
 		<link rel="stylesheet" href="css/error.css" />
 	</head>
 	<body>
 		<div class="message">알 수 없는 오류입니다.</div>
-		<img src="css/img/warning.png" class="warning" alt="warning" width="200" />
+		<img src="img/warning.png" class="warning" alt="warning" width="200" />
 	</body>
 </html>
