+++ public/index.html
@@ -5,7 +5,7 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<meta http-equiv="X-UA-Compatible" content="ie=edge" />
 		<title>TODOS</title>
-		<link rel="shortcut icon" href="css/img/favicon.ico">
+		<link rel="shortcut icon" href="img/favicon.ico">
 		<link rel="stylesheet" href="css/todos.css" />
 		<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
 	</head>
@@ -22,7 +22,7 @@
 		<section id="todos-section">
 			<div class="todo todos">
 				<div class="todo todos-title">
-					todo<img src="css/img/trash-can.png" class="trash-can" title="휴지통" alt="trash-can" width="20">
+					todo<img src="img/trash-can.png" class="trash-can" title="휴지통" alt="trash-can" width="20">
 					<hr class="todo todos-underline">
 				</div>
 				<div id="todo" class="todo dropArea">
@@ -31,7 +31,7 @@
 
 			<div class="doing todos">
 				<div class="doing todos-title">
-					doing<img src="css/img/trash-can.png" class="trash-can" title="휴지통" alt="trash-can" width="20">
+					doing<img src="img/trash-can.png" class="trash-can" title="휴지통" alt="trash-can" width="20">
 					<hr class="doing todos-underline">
 				</div>
 				<div id="doing" class="doing dropArea">
@@ -40,7 +40,7 @@
 
 			<div class="done todos">
 				<div class="done todos-title">
-					done<img src="css/img/trash-can.png" class="trash-can" title="휴지통" alt="trash-can" width="20">
+					done<img src="img/trash-can.png" class="trash-can" title="휴지통" alt="trash-can" width="20">
 					<hr class="done todos-underline">
 				</div>
 				<div id="done" class="done dropArea">
