+++ public/js/sign-up.js
