+++ public/js/todos.js
@@ -1,4 +1,4 @@
-const todosDB = require('./todosDB.js');
+const todosDB = require('../../todosDB.js');
 
 const getTodosList = user_id => {
 	const todosList = todosDB.getTodosList(user_id);
